requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
pydantic-settings
python-dotenv
python-jose[cryptography]
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
httpx
google-genai
google-generativeai
slowapi
python-multipart
email-validator
reportlab
pypdf
pillow
requests
aiohttp
edge-tts
python-jose
pdfplumber
iyzipay
PyJWT
# ── Yerel Kamera & Video Analiz ──────────────────
opencv-python-headless
ultralytics
# Not: FFmpeg sistem düzeyinde kurulmalı (apt/brew/winget)
